+++ clikt/src/main/kotlin/com/github/ajalt/clikt/parameters/types/choice.kt
@@ -1,5 +1,6 @@
 package com.github.ajalt.clikt.parameters.types
 
+import com.github.ajalt.clikt.completion.CompletionCandidates
 import com.github.ajalt.clikt.parameters.arguments.ProcessedArgument
 import com.github.ajalt.clikt.parameters.arguments.RawArgument
 import com.github.ajalt.clikt.parameters.arguments.convert
@@ -29,6 +30,7 @@ private fun errorMessage(choice: String, choices: Map<String, *>): String {
 fun <T : Any> RawArgument.choice(choices: Map<String, T>): ProcessedArgument<T, T> {
     require(choices.isNotEmpty()) { "Must specify at least one choice" }
     return convert { choices[it] ?: fail(errorMessage(it, choices)) }
+            .run { copy(transformValue, transformAll, completionCandidates = CompletionCandidates.Fixed(choices.keys)) }
 }
 
 /**
@@ -71,7 +73,9 @@ fun RawArgument.choice(vararg choices: String): ProcessedArgument<String, String
 fun <T : Any> RawOption.choice(choices: Map<String, T>,
                                metavar: String = mvar(choices.keys)): NullableOption<T, T> {
     require(choices.isNotEmpty()) { "Must specify at least one choice" }
-    return convert(metavar) { choices[it] ?: fail(errorMessage(it, choices)) }
+    return convert(metavar, completionCandidates = CompletionCandidates.Fixed(choices.keys)) {
+        choices[it] ?: fail(errorMessage(it, choices))
+    }
 }
 
 /**